# actions/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "beads-bridge tool bridge",
    "version": "1.1.0",
    "description": "Structured read/write actions over Trillium's federated bead stores (brain, task, review, etc.). Every bead id encodes its store prefix (e.g. review-fmt). Read actions return the bead plus its available_actions so the model can act without navigation. Mutations return the canonical post-state, not bare text."
  },
  "servers": [
    {
      "url": "__FUNNEL_BASE__"
    }
  ],
  "paths": {
...
    "/action/comment": {
      "post": {
        "operationId": "comment",
        "summary": "Add a comment to a bead.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Canonical post-state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bead"
                }
              }
            }
          }
        }
      }
    },
...
    "/action/approve": {
      "post": {
        "operationId": "approve",
        "summary": "Approve an OPEN bead: comment + close. 409 if not open.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BeadBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Canonical post-state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bead"
                }
              }
            }
          },
          "409": {
            "description": "Bead not open"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BeadBody": {
        "type": "object",
        "required": [
          "bead_id"
        ],
        "properties": {
          "bead_id": {
            "type": "string",
            "example": "review-fmt"
          },
          "idempotency_key": {
            "type": "string",
            "description": "Optional; replays return the stored response"
          }
        }
      },
      "CommentBody": {
        "type": "object",
        "required": [
          "bead_id",
          "text"
        ],
        "properties": {
          "bead_id": {
            "type": "string",
            "example": "review-fmt"
          },
          "text": {
            "type": "string",
            "description": "Comment or note text"
          },
          "idempotency_key": {
            "type": "string",
            "description": "Optional; replays return the stored response"
          }
        }
      },
      "Bead": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "bead_id": {
            "type": "string"
          },
          "store": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "enum": [
              "open",
              "closed"
            ]
          },
          "priority": {
            "type": "string"
          },
          "labels": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "comments": {
            "type": "string"
          },
          "available_actions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "wrote": {
            "type": "string",
            "description": "Present on mutations"
          },
          "transitioned": {
            "type": "string",
            "description": "Present on approve/reject/done"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}