# src/routes/read.ts
// Read routes: /, /help, /next, /:store, /:id. Mounted after the resume
// router so literal /fetch/* and /resume win over the param routes.
import { Router, Request, Response, NextFunction } from 'express'
import { execSync } from 'child_process'
import { readFileSync } from 'fs'
import path from 'path'
import { BASE, STORES, storeAbout, fillTokens } from '../config'
import { qstr, pstr, bd, storeFromId, cacheTag } from '../util'
import { wrap } from '../wrap'
import { isWebAgent } from '../agent-detect'

export const mountOrder = -10
export const readRouter = Router()

// GET / — the research-assistant prompt (markdown), served verbatim from prompt.md.
// The cache-tag placeholder is minted here so the URL ChatGPT receives is literal.
// GET / — agents get prompt.md (the bridge action brief); browsers get the
// SPA (Vite build output, file-based routes in web/src/routes/).
readRouter.get('/', (req: Request, res: Response) => {
  if (!isWebAgent(req.get('user-agent'), req.get('accept')))
    return void res.sendFile(path.join(__dirname, '..', '..', 'web', 'dist', 'index.html'))
  try {
    const prompt = fillTokens(readFileSync(path.join(__dirname, '..', '..', 'prompt.md'), 'utf8'))
      .replace('<server-issued-tag>', cacheTag())
    res.type('text/markdown').send(prompt)
  } catch {
    res.type('text/plain').send('# Beads Bridge\n\nMissing prompt.md. See /help for the action reference.')
  }
})

// GET /help — the bridge action reference
readRouter.get('/help', (_req: Request, res: Response) => {
  const lines: string[] = [
    '# Beads Bridge — Actions',
    '',
    'A GET-only, plain-text bridge to Trillium\'s federated bead stores (brain, task, review, etc.).',
    'Every bead has a stable id like `review-3y3`; its store is derived from the id prefix.',
    '',
    '## Instructions',
    '',
    'If you are an assistant handling this queue, start by fetching exactly this URL:',
    '',
    `    ${BASE}/next?cache=${cacheTag()}`,
    '',
    'Read the response, then follow the bead\'s instructions exactly using the GET URLs it lists.',
    'The `cache=` tag is a unique per-response value generated by the server — always fetch the',
    'URL exactly as given, including the tag. Never rely on a result already in this conversation:',
    'the queue changes and cached bodies are stale.',
    '',
    '## Actions',
    `  GET ${BASE}/next?cache=<server-issued-tag>  — next human-queued bead, with its actions`,
    `  GET ${BASE}/{bead-id}                       — view a bead (e.g. /review-3y3)`,
    `  GET ${BASE}/{bead-id}/approve               — approve + close (no payload)`,
    `  GET ${BASE}/{bead-id}/reject                — reject, stays open (no payload)`,
    `  GET ${BASE}/{bead-id}/done                  — mark handled + close (no payload)`,
    `  GET ${BASE}/{bead-id}/comment?text=...      — add a comment (or use ?t=...)`,
    `  GET ${BASE}/{bead-id}/note?text=...         — append a note (or use ?t=...)`,
    `  GET ${BASE}/{bead-id}/label?add=triaged    — acknowledged/triaged; hidden from /next`,
    `  GET ${BASE}/{bead-id}/label?add=ready       — signal decision made; downstream may act`,
    `  GET ${BASE}/{bead-id}/close                 — close a bead`,
    `  GET ${BASE}/{bead-id}/label?add=...         — change labels`,
    `  GET ${BASE}/{store}?q=...                   — search a store`,
    `  GET ${BASE}/fetch/{resume}/unconfirmed     — resume triage: pending bullets + workExperience`,
    `  GET ${BASE}/fetch/{resume}/complete        — full resume markdown with green/orange ledger`,
    `  GET ${BASE}/fetch/{resume}/job-description — posting job bead verbatim (e.g. /fetch/resumes-zak/complete)`,
    '',
    '## Decisions and consumption',
    '',
    'Each decision is recorded on the bead (comment/label), then a consumer acts.',
    '`triaged` = someone looked; the bead is hidden from /next but stays open.',
    '`ready` = decision made; downstream may execute (e.g. a `ready`-labelled PR review',
    'is merged by the fleet poller), then the bead is closed.',
    '',
    '## If a text URL is rejected',
    '',
    'If a `?text=...` URL is rejected by the web-fetch safety filter, retry the SAME text',
    'using the short form `?t=...`. If that is rejected too, do not block on it: record the',
    'decision with the no-payload verb (approve/reject/done) and state the reasoning in chat.',
  ]
  res.type('text/plain').send(lines.join('\n'))
})

// GET /next — next open bead labeled `human`, priority order.
// The `review` store is the human-decision queue, so it is served FIRST.
readRouter.get('/next', (_req: Request, res: Response) => {
  const queueOrder = ['review', ...STORES.filter(s => s !== 'review')]
  const priorities = ['P0', 'P1', 'P2', 'P3', 'P4']
  for (const store of queueOrder) {
    for (const prio of priorities) {
      try {
        const raw = execSync(
          `${store} list --label human --state open --priority ${prio} --limit 1 --exclude-label triaged --exclude-label ready 2>/dev/null`,
          { encoding: 'utf8', timeout: 8000 }
        ).trim()
        if (!raw) continue

        const match = raw.match(/\b([a-z_]+-[a-z0-9]+)\b/)
        if (!match) continue
        const id = match[1]

        const body = bd(store, `show ${id}`)
        const comments = bd(store, `comments ${id}`)
        const full = comments ? `${body}\n\n## Comments\n${comments}` : body
        const directions =
          'YOUR ROLE: research this item as far as you reasonably can ' +
          '(read the bead, follow links such as the PR/repo/checks) and present a ' +
          'short briefing to the user: what it is, what you verified, what decision ' +
          'or action it is asking for, and what is missing.\n' +
          'READ-ONLY: do not call the write URLs below unless the user explicitly asks.'

        return res.type('text/plain').send(wrap({
          title: `Next: ${id} (${prio})`,
          body: `${directions}\n\n${full}`,
          meta: { store, id },
          actions: [
            `GET ${BASE}/${id}            — view bead (read)`,
            `GET ${BASE}/${id}/approve    — approve; comment + close`,
            `GET ${BASE}/${id}/reject     — reject; comment, stays open`,
            `GET ${BASE}/${id}/done       — mark handled; comment + close`,
          ]
        }))
      } catch {
        continue
      }
    }
  }

  res.type('text/plain').send(
    '# No queued items\n\nNo open beads with the `human` label found.\n\nFetch / to browse stores manually.'
  )
})

// GET /{store} — list or search store
readRouter.get('/:store', (req: Request, res: Response, next: NextFunction) => {
  const store = pstr(req.params.store)
  if (!STORES.includes(store)) return next()

  const q = qstr(req.query.q)
  const raw = q
    ? bd(store, `search "${q}"`)
    : bd(store, 'list')

  const about = storeAbout(store)
  res.type('text/plain').send(wrap({
    title: `${store} — ${about}`,
    body: raw || '(empty)',
    meta: { store },
    actions: [
      `GET /${store}?q=keyword   — search this store`,
      `GET /{bead-id}            — view a specific bead`,
      `GET ${BASE}/next?cache=<server-issued-tag> — next human-queued item`,
    ]
  }))
})

// GET /beads/{id1+id2+...} — show multiple beads on one page (max 20).
// Lets one literal URL carry the whole deeper-context roster.
readRouter.get('/beads/:ids', (req: Request, res: Response) => {
  const raw = String(pstr(req.params.ids)).split('+').map((x: string) => x.trim()).filter(Boolean).slice(0, 40)
  const valid = raw.filter((id: string) => /^[A-Za-z][A-Za-z0-9_-]{2,64}$/.test(id) && storeFromId(id))
  if (!valid.length) return res.type('text/plain').status(400).send('No valid bead ids in URL.')
  const parts = valid.map((id: string) => {
    const store = storeFromId(id)!
    const body = bd(store, `show ${id}`)
    const comments = bd(store, `comments ${id}`)
    return [`# ${id} (STORE: ${store})`, '', comments ? `${body}\n\n## Comments\n${comments}` : body].join('\n')
  })
  res.type('text/plain').send(wrap({
    title: `Beads: ${valid.join(' + ')}`,
    body: parts.join('\n\n---\n\n'),
    actions: valid.map(id => `GET ${BASE}/${id} — view ${id} alone`),
  }))
})

// GET /{bead-id} — show bead
readRouter.get('/:id', (req: Request, res: Response, next: NextFunction) => {
  const id = pstr(req.params.id)
  const store = storeFromId(id)
  if (!store) return next()

  const body = bd(store, `show ${id}`)
  const comments = bd(store, `comments ${id}`)
  const full = comments ? `${body}\n\n## Comments\n${comments}` : body

  res.type('text/plain').send(wrap({
    title: id,
    body: full,
    meta: { store, id },
    actions: [
      `GET ${BASE}/${id}/comment?text=...    — add a comment (or use ?t=...)`,
      `GET ${BASE}/${id}/note?text=...       — append a note (or use ?t=...)`,
      `GET ${BASE}/${id}/approve             — approve; comment + close`,
      `GET ${BASE}/${id}/reject              — reject; comment, stays open`,
      `GET ${BASE}/${id}/done                — mark handled; comment + close`,
      `GET ${BASE}/${id}/close               — close without comment`,
    ]
  }))
})